Creacion de API usando spring web(Tomcat), postgresql(symplificadb create), lombok(gestion de db(CRUD)), Spring Boot JPA(hibernate), Creacion de CRUD, Implementacion de Excepciones(usando H2 mientras se verifican problemas de comunicacion con la base de datos)

+++ demoSymplifica/src/main/java/com/symplifica/demoSymplifica/services/ProductServiceImplementation.java
@@ -1,13 +1,16 @@
 package com.symplifica.demoSymplifica.services;
 
 import com.symplifica.demoSymplifica.entity.Product;
+import com.symplifica.demoSymplifica.error.ProductNotFoundException;
 import com.symplifica.demoSymplifica.repository.ProductRepository;
-import jakarta.annotation.Nullable;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.method.annotation.HandlerMethodValidationException;
 
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
+
 
 @Service
 public class ProductServiceImplementation implements ProductService{
@@ -15,12 +18,12 @@ public class ProductServiceImplementation implements ProductService{
     ProductRepository productRepository;
 
     @Override
-    public Product saveProduct(Product product) {
+    public Product saveProduct(Product product){
         return productRepository.save(product);
     }
 
     @Override
-    public Product updateProduct(Long id, Product product) {
+    public Product updateProduct(Long id, Product product) throws ProductNotFoundException{
         Product ofTheProduct = productRepository.findById(id).get();
         if(Objects.nonNull(product.getName()) && !"".equalsIgnoreCase(product.getName())){
             ofTheProduct.setName(product.getName());
@@ -39,12 +42,29 @@ public class ProductServiceImplementation implements ProductService{
     }
 
     @Override
-    public void deleteProduct(Long id) {
+    public void deleteProduct(Long id) throws ProductNotFoundException{
+        Optional<Product> product = productRepository.findById(id);
+        
+        if(!product.isPresent()){
+            throw  new ProductNotFoundException("The product with id: "+ id +", doesn't exist in the inventary");
+        }
         productRepository.deleteById(id);
     }
 
     @Override
-    public List<Product> listAllProducts() {
+    public Product findProductById(Long id) throws ProductNotFoundException{
+        return productRepository.findById(id).get();
+    }
+
+    @Override
+    public Product findProductByName(String name) throws ProductNotFoundException{
+        return productRepository.findByName(name);
+    }
+
+    @Override
+    public List<Product> listAllProducts() throws ProductNotFoundException {
         return productRepository.findAll();
     }
+
+
 }
